+++ test/dndManager.js
@@ -1,8 +1,7 @@
 const chai = require('chai')
 const path = require('path')
 const DndManager = require('../app/utils/dndManager')
-const Settings = require('./../app/utils/settings')
-const testSettingsLocaction = path.join(__dirname, '/test-settings.json')
+const Store = require('electron-store')
 
 chai.should()
 
@@ -11,7 +10,11 @@ describe('dndManager', function () {
   let dndManager = null
 
   beforeEach(() => {
-    settings = new Settings(testSettingsLocaction)
+    settings = new Store({
+      cwd: path.join(__dirname),
+      name: 'test-settings',
+      defaults: require('../app/utils/defaultSettings')
+    })
     dndManager = new DndManager(settings)
   })
 
@@ -56,7 +59,7 @@ describe('dndManager', function () {
     dndManager = null
 
     if (settings && settings.destroy) {
-      settings.destroy()
+      require('fs').unlink(path.join(__dirname, '/test-settings.json'), (_) => {})
       settings = null
     }
   })